requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
Django==4.0.6
django-bootstrap3==22.1
django-bootstrap4==22.2
djangorestframework==3.14.0
Pillow==9.2.0
psycopg2==2.9.3
pytz==2022.4
soupsieve==2.3.2.post1
sqlparse==0.4.2
